test: split TriggerTest by subject class
Move ClickTrigger-specific tests to ClickTriggerTest and the
DomEventTrigger install test to DomEventTriggerTest, leaving
TriggerTest with the generic Trigger contract tests only.
